+++ src/Application/Orders/Queries/GetOrderByFormIds/OrderCustomer.cs
@@ -6,7 +6,7 @@ public class OrderCustomer
     public long CitizenshipNumber { get; set; }
     public string First { get; set; }
     public string Last { get; set; }
-    public int CustomerNumber { get; set; }
+    public UInt64 CustomerNumber { get; set; }
     public string BranchCode { get; set; }
     public string BusinessLine { get; set; }
 }
