+++ workers/Worker.App/Models/SendMailTemplateRequest.cs
@@ -5,7 +5,7 @@ public class SendMailTemplateRequest
     public HeaderInfo headerInfo { get; set; }
     public string templateParams { get; set; }
     public string template { get; set; }
-    public int customerNo { get; set; }
+    public UInt64 customerNo { get; set; }
     public string email { get; set; }
     public Process process { get; set; }
 }
@@ -29,7 +29,7 @@ public class SendMailTemplateHeaderInfo
 
 public class EmailTemplateParams
 {
-    public int MusteriNo { get; set; }
+    public UInt64 MusteriNo { get; set; }
     public string MusteriAdSoyad { get; set; }
     public string Title { get; set; }
 }
